+++ commands/utility/intro.js
@@ -1,4 +1,4 @@
-const { SlashCommandBuilder, EmbedBuilder, AttachmentBuilder } = require('discord.js');
+const { SlashCommandBuilder, EmbedBuilder } = require('discord.js');
 
 function getColorHex(colorName) {
 	const isHexCode = /^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$/.test(colorName);
@@ -34,13 +34,11 @@ module.exports = {
 				.setRequired(false)),
 	async execute(interaction) {
 		const user = interaction.user;
-		const filename = `avatar.${user.avatar?.startsWith('a_') ? 'gif' : 'png'}`;
-		const avatar = new AttachmentBuilder(user.displayAvatarURL(), { name: filename });
 		const name = interaction.options.getString('name');
 		const hobby = interaction.options.getString('hobby') || '-';
 		const location = interaction.options.getString('location') || '-';
 		const description = interaction.options.getString('description');
-		const colorInput = interaction.options.getString('color') || '#0099ff';
+		const colorInput = interaction.options.getString('color') || '#9B7AAC';
 		const color = getColorHex(colorInput);
 		if (!color) {
 			await interaction.reply({ content: 'Sorry masbro, Warnanya tidak valid. Pastikan kamu menggunakan [kode hex](https://www.google.com/search?q=Pemilih+warna) yang valid. Contoh: #123ABC, atau kosongkan saja kalau mau pakai warna default', ephemeral: true });
@@ -50,9 +48,8 @@ module.exports = {
 		const introEmbed = new EmbedBuilder()
 			.setColor(color)
 			.setTitle('Perkenalkan')
-			.setThumbnail(`attachment://${filename}`)
+			.setThumbnail(user.displayAvatarURL())
 			.setDescription(`Nama: ${name}\nHobi: ${hobby}\nDaerah asal: ${location}\n\n${description}`);
-
-		await interaction.reply({ embeds: [introEmbed], files: [avatar] });
+		await interaction.reply({ embeds: [introEmbed] });
 	},
 };