+++ events/ready.js
@@ -3,7 +3,7 @@ const { onlineChannelId } = require('../config.json');
 
 let currentMode;
 if (process.env.launchType === 'Server') {
-	currentMode = 'Server / Pylexnode Server - Yozora berjalan dengan stabil';
+	currentMode = 'Server - Yozora berjalan dengan stabil';
 }
 else {
 	currentMode = 'Unknown - ada masalah pada peluncuran Yozora. Beberapa fungsi system dimatikan';
@@ -13,11 +13,11 @@ module.exports = {
 	name: Events.ClientReady,
 	once: true,
 	execute(client) {
-		console.log(`Ready! Logged in as ${client.user.tag}`);
+		console.log("\x1b[33m%s\x1b[0m", `Logged in as ${client.user.tag} on ${process.env.launchType}`);
 		if (process.env.launchType !== 'Local') {
 			const channel = client.channels.cache.get(onlineChannelId);
 			if (channel) {
-				channel.send(`<a:rikomunity:1208619104924340276>・Yozora online dalam mode ${currentMode}`);
+				channel.send(`<a:zrikomunity:1218433304030613524> ・Yozora online dalam mode ${currentMode}`);
 			}
 			else {
 				console.error(`Invalid channel ID: ${onlineChannelId}`);